Polish translation – corrections
Hello, this is my first contribution of GitHub so I hope I did everything correctly.

I decided to come up with some corrections to the polish translation primarily because of the translation of "clear" in the current "values" section – a literal translation of the current "wyczyść" would something like "clean it up" and is very confusing. For example, in case of water leakage sensor a new user or one without experience would rather understand "wyczyść" as an indication that there is water leakage to clean, when in fact it's a translation of state of there not being any leakage at all. Also, "tampered" as "manipulacja" has no meaning – it even leaves no clue what can be happening with the device.

Changes described above are the most important and I think should be a priority. The rest that I suggested are mostly to make it sound more natural. Additionally, in "ota" and "settingsSchemaDescriptions" I've come up with changes to make the translations more informative and easier to understand (based on the reference file https://github.com/Nerivec/zigbee2mqtt-windfront/blob/main/src/i18n/locales/en.json ). I also unified the abbreviation for second where applicable.

I hope it will come as helpful. Should there be any questions, I'll try to answer as soon as possible.